Split normalizers into a normalizer mod
